translate address

+++ mmu.h
@@ -81,7 +81,7 @@ int second_chance_replacement();
 void handle_page_fault(PCB *pcb, int process_Count ,int process_id, int virtual_page, char readwrite_flag);
 void load_page_from_disk(int process_id, int virtual_page, int physical_page);
 void swap_page_to_disk(int process_id, int virtual_page, int physical_page);
-long translate_address(int process_id, long virtual_address, PCB* pcb, char rw_flag);
+int translate_address(int process_id, long virtual_address, PCB* pcb, char rw_flag);
 void update_disk_operations(int current_time);
 bool is_page_in_disk_queue(int process_id, int virtual_page);
 void print_memory_log(const char* format, ...);
